Add functionality, too.
